matter header width
