+++ src/images/gatsby-astronaut.png
