+++ src/images/gatsby-icon.png
